refactor: convert object API to async/await (#2666)
* refactor: convert object API to async/await

* fix: object export
