+++ src/core/components/index.js
@@ -9,6 +9,20 @@ exports.bitswap = {
 exports.config = require('./config')
 exports.id = require('./id')
 exports.init = require('./init')
+exports.object = {
+  data: require('./object/data'),
+  get: require('./object/get'),
+  links: require('./object/links'),
+  new: require('./object/new'),
+  patch: {
+    addLink: require('./object/patch/add-link'),
+    appendData: require('./object/patch/append-data'),
+    rmLink: require('./object/patch/rm-link'),
+    setData: require('./object/patch/set-data')
+  },
+  put: require('./object/put'),
+  stat: require('./object/stat')
+}
 exports.ping = require('./ping')
 exports.start = require('./start')
 exports.stop = require('./stop')
@@ -17,6 +31,5 @@ exports.version = require('./version')
 exports.legacy = { // TODO: these will be removed as the new API is completed
   dag: require('./dag'),
   libp2p: require('./libp2p'),
-  object: require('./object'),
   pin: require('./pin')
 }
