+++ src/core/components/init.js
@@ -102,7 +102,20 @@ module.exports = ({
     })
 
     const dag = Commands.legacy.dag({ _ipld: ipld, _preload: preload })
-    const object = Commands.legacy.object({ _ipld: ipld, _preload: preload, dag, _gcLock: gcLock })
+    const object = {
+      data: Commands.object.data({ ipld, preload }),
+      get: Commands.object.get({ ipld, preload }),
+      links: Commands.object.links({ dag }),
+      new: Commands.object.new({ ipld, preload }),
+      patch: {
+        addLink: Commands.object.patch.addLink({ ipld, gcLock, preload }),
+        appendData: Commands.object.patch.appendData({ ipld, gcLock, preload }),
+        rmLink: Commands.object.patch.rmLink({ ipld, gcLock, preload }),
+        setData: Commands.object.patch.setData({ ipld, gcLock, preload })
+      },
+      put: Commands.object.put({ ipld, gcLock, preload }),
+      stat: Commands.object.stat({ ipld, preload })
+    }
 
     const pinManager = new PinManager(repo, dag)
     await pinManager.load()
@@ -134,6 +147,7 @@ module.exports = ({
       initOptions: options,
       ipld,
       keychain,
+      object,
       peerInfo,
       pinManager,
       preload,
@@ -276,6 +290,7 @@ function createApi ({
   initOptions,
   ipld,
   keychain,
+  object,
   peerInfo,
   pinManager,
   preload,
@@ -302,6 +317,7 @@ function createApi ({
     config: Commands.config({ repo }),
     id: Commands.id({ peerInfo }),
     init: () => { throw new AlreadyInitializedError() },
+    object,
     start,
     version: Commands.version({ repo })
   }
