+++ src/core/components/object/patch/rm-link.js
@@ -1,0 +1,12 @@
+'use strict'
+
+module.exports = ({ ipld, gcLock, preload }) => {
+  const get = require('../get')({ ipld, preload })
+  const put = require('../put')({ ipld, gcLock, preload })
+
+  return async function rmLink (multihash, linkRef, options) {
+    const node = await get(multihash, options)
+    node.rmLink(linkRef.Name || linkRef.name)
+    return put(node, options)
+  }
+}
