+++ src/core/components/object/stat.js
@@ -1,0 +1,28 @@
+'use strict'
+
+const dagPB = require('ipld-dag-pb')
+
+module.exports = ({ ipld, preload }) => {
+  const get = require('./get')({ ipld, preload })
+  return async function stat (multihash, options) {
+    options = options || {}
+
+    const node = await get(multihash, options)
+    const serialized = dagPB.util.serialize(node)
+    const cid = await dagPB.util.cid(serialized, {
+      cidVersion: 0
+    })
+
+    const blockSize = serialized.length
+    const linkLength = node.Links.reduce((a, l) => a + l.Tsize, 0)
+
+    return {
+      Hash: cid.toBaseEncodedString(),
+      NumLinks: node.Links.length,
+      BlockSize: blockSize,
+      LinksSize: blockSize - node.Data.length,
+      DataSize: node.Data.length,
+      CumulativeSize: blockSize + linkLength
+    }
+  }
+}
